Alpha 1. Fix display of non-aligned row at end of page.
Imagine a phrase with delai_ui = 6. We display 33 rows,
so next page (page 1) would start at time 6*33 = 198.
Now, if a row is defined at time 196, it cannot be displayed in page 0,
as we need 5 lines for a row, and the viewport height is 200.
But it won't be displayed either in page 1, since 196 < 198.

It also triggered an assertion failure when coming from top of page 1,
after cursor up: time cursor went from 198 to 196, switching to page 0.
Since 196 is out of viewport, a new page switch was initiated.
This double switch was detected (assert failure).

Solution: Switch page by 32 rows instead of 33, to leave some margin for
this case. Plus, it is better ui-wise, since bars remains aligned.